rename log -> serverLogger
