+++ packages/cli/src/commands/__test__/init.test.ts
@@ -1,7 +1,7 @@
 import prompts from "prompts";
 import fsExtra, { removeSync } from "fs-extra";
 import { handler, InitArguments } from "../init";
-import { log } from "../../util/log";
+import { log } from "../../util/serverLogger";
 
 jest.useFakeTimers();
 jest.mock("../../util/log");
