+++ packages/cli/src/util/config/__test__/config.test.ts
@@ -1,4 +1,4 @@
-import { log } from "../../log";
+import { log } from "../../serverLogger";
 import fsExtra from "fs-extra";
 import crypto from "crypto";
 import {
